% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decTTest.R
\name{logAdjT}
\alias{logAdjT}
\title{log-adjusted-t
}
\usage{
logAdjT(x, y)
}
\arguments{
\item{x}{A numeric vector representing sample 1.}

\item{y}{A numeric vector representing sample 2.}
}
\value{
A list containing the original p-value and the adjusted p-value.
}
\description{
This function performs a standard t-test and an adjusted version
with a log-transformed correction for small sample sizes.
}
