% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decTTest.R
\name{t_dec}
\alias{t_dec}
\title{Decremented t-Test with Log-Transformed Correction}
\usage{
t_dec(x, y)
}
\arguments{
\item{x}{A numeric vector representing sample 1.}

\item{y}{A numeric vector representing sample 2.}
}
\value{
A list containing the original p-value and the corrected p-value.
}
\description{
This function performs a standard t-test and an adjusted version
with a log-transformed correction for small sample sizes.
}
